// packages/base-tool/src/enums.ts
/* eslint-disable no-unused-vars */
export enum Icons {
	POINT_DOWN = 'point_down',
	POINT_UP = 'point_up',
	POINT_LEFT = 'point_left',
	POINT_RIGHT = 'point_right',
	SETTINGS = 'settings',
	RETRY = 'retry',
	PLAY_BUTTON = 'play_button',
	MINUS = 'minus',
	PLUS = 'plus',
	RANDOMIZE = 'randomize',
	TRASH = 'trash',
	VIEW = 'view',
	VIEW_ALT = 'view-alt',
	HIDE = 'hide',
	HIDE_ALT = 'hide-alt',
	CLEAR_INPUT = 'clear_input',
	COPY_PASTE = 'copy_paste',
	QR_CODE = 'qr-code',
	DICE_THROW = 'dice_throw',
	ADD = 'add',
	CALCULATE = 'calculate',
	SORT = 'sort',
}

export enum Environments {
	TOOL = 'tool',
	SETTINGS = 'settings',
}

export enum ButtonSizes {
	SMALL = 'small',
	MEDIUM = 'medium',
	LARGE = 'large',
}

export enum ButtonColors {
	PRIMARY = 'primary',
	SECONDARY = 'secondary',
	LIGHT = 'light',
	SHUTTLE_GRAY = 'shuttle-gray',
	SOLITUDE = 'solitude',
	NEGATIVE = 'negative',
}

export enum ButtonOrientations {
	HORIZONTAL = 'horizontal',
	VERTICAL = 'vertical',
}

export enum LoadingStatus {
	NONE = 'none',
	LOADING = 'loading',
	LOADED = 'loaded',
}

export enum ToolTipPlacement {
	BOTTOM = 'bottom',
	TOP = 'top',
}

export enum DropdownPositions {
	AUTO = 'auto',
	TOP = 'top',
	BOTTOM = 'bottom',
}